refactor(dashboard): consolidate time formatting into src/lib/format-time.ts
Unify 8 duplicate time functions (relativeTime, formatElapsed, formatDuration,
formatTimeAgo, timeAgo) from 7 files into a single module.

- relativeTime: configurable fallback ('정보 없음' default, '방금' for mission)
- formatElapsed: Korean output (초/분/시간)
- formatElapsedCompact: English compact (s/m/h)
- formatDuration: Korean with day support
- formatTimeAgo: handles string/number timestamps with unix auto-detection

Net: -4 lines (78 removed, 74 added). Zero behavior change.
