+++ dashboard/src/components/goals.ts
@@ -17,6 +17,7 @@ import {
   tasksByStatus,
 } from '../store'
 import type { Goal, MdalLoop, Task } from '../types'
+import { formatElapsedCompact } from '../lib/format-time'
 
 // -- Filter state ------------------------------------------------
 
@@ -93,12 +94,6 @@ function horizonColor(h: string): string {
     case 'long': return '#818cf8'
     default: return '#888'
   }
-}
-
-function formatElapsed(seconds: number): string {
-  if (seconds < 60) return `${Math.round(seconds)}s`
-  if (seconds < 3600) return `${Math.floor(seconds / 60)}m ${Math.round(seconds % 60)}s`
-  return `${Math.floor(seconds / 3600)}h ${Math.floor((seconds % 3600) / 60)}m`
 }
 
 function formatMetric(value: number): string {
@@ -281,7 +276,7 @@ function LoopRow({ loop }: { loop: MdalLoop }) {
           <span class=${formatMetricDelta(loop).startsWith('+') ? 'planning-loop-good' : 'planning-loop-bad'}>
             Delta ${formatMetricDelta(loop)}
           </span>
-          <span>Elapsed ${formatElapsed(loop.elapsed_seconds)}</span>
+          <span>Elapsed ${formatElapsedCompact(loop.elapsed_seconds)}</span>
         </div>
 
         <div class="planning-loop-target">${loop.target || '명시된 목표가 없습니다'}</div>
